+++ packages/core/src/utils.ts
@@ -1,0 +1,8 @@
+import type { Derived } from './derived.js';
+import type { Signal } from './signal.js';
+
+export function validate(v: Signal<unknown> | Derived<unknown>) {
+  if ('validate' in v) {
+    v.validate();
+  }
+}
