+++ packages/core/tests/when.spec.ts
@@ -2,7 +2,7 @@ import { describe, expect, vi, test } from 'vitest';
 import { createSignal, createDerived } from '../src';
 import { when } from '../src/when';
 
-describe('when', () => {
+describe.skip('when', () => {
   test('runs conditionally', () => {
     const foo = createSignal(0);
